requirements.txt
# Telegram Bot
aiogram>=3.0.0

# Database
sqlalchemy>=2.0.0
aiosqlite>=0.19.0

# Parsing - Camoufox
camoufox>=0.4.0

# Parsing - Selenium
selenium>=4.40.0
selenium-stealth>=1.0.0
webdriver-manager>=4.0.0

# AI
httpx>=0.25.0
requests>=2.31.0

# Scheduler
apscheduler>=3.10.0

# Utils
python-dotenv>=1.0.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
aiofiles>=23.0.0
Pillow>=10.0.0  # Для сжатия скриншотов
aiohttp>=3.9.0  # Для загрузки фото товаров
psutil>=5.9.0  # Для очистки процессов
